# api/internal/server/docs/swagger.yaml
basePath: /
definitions:
  auth.Principal:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      picture_url:
        type: string
      type:
        $ref: '#/definitions/auth.PrincipalType'
    type: object
  auth.PrincipalType:
    enum:
    - user
    - service_account
    - anyone
    - anonymous
    type: string
    x-enum-varnames:
    - PrincipalUser
    - PrincipalServiceAccount
    - PrincipalAnyone
    - PrincipalAnonymous
...
  repo.Grant:
    properties:
      document_id:
        type: string
      granted_by:
        $ref: '#/definitions/auth.Principal'
      id:
        type: string
      principal:
        $ref: '#/definitions/auth.Principal'
      role:
        $ref: '#/definitions/repo.Role'
    type: object
...
  repo.Role:
    enum:
    - ""
    - viewer
    - commenter
    - editor
    - owner
    type: string
    x-enum-varnames:
    - RoleNone
    - RoleViewer
    - RoleCommenter
    - RoleEditor
    - RoleOwner
...
info:
  contact: {}
  description: This API powers a review tool where people can upload one self-contained
    HTML file, view it, and leave comments anchored to specific text ranges inside
    that HTML.
  title: aidocs API
  version: v1
paths:
...
  /v1/documents/{id}/grants:
    get:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: List document grants
      tags:
      - grants
    post:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Create document grant
      tags:
      - grants
  /v1/documents/{id}/grants/{grant_id}:
    delete:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Grant ID
        in: path
        name: grant_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Delete document grant
      tags:
      - grants
    patch:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Grant ID
        in: path
        name: grant_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Grant'
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Update document grant
      tags:
      - grants
...
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
  cookieAuth:
    in: cookie
    name: aidocs_session
    type: apiKey
swagger: "2.0"